Add files via upload
This code displays amount of ingredients in the machine. Allows user to select type of coffee he/she wants to order . It receives money in form of notes and calculates total money received and then checks it with cost of coffee ordered and if excess money is received refunds the balance amount and then processes the order and updates the resources repository.